:art: Update app layout and dependencies
This update consists of changes in the application's layout and components in the src/App.tsx file. It also includes an upgrade in package dependencies, removal of some old dependencies, and addition of new ones, such as "react-d3-cloud", "react-youtube", and "tailwindcss-animated". Lastly, it includes the installation of a babel plugin, "@babel/plugin-proposal-private-property-in-object".