:green_heart: Add step to install sync issues in build workflow
A new step has been added to the GitHub Actions build workflow. This step runs the `sync-issues.js` script, using a secret TOKEN, to enable the syncing of issues as part of the build process. This change ensures that the issues are always up-to-date to reflect the current status of the project.