:art: Update asset-manifest.json and articles content
The asset-manifest.json has been updated with the new js file names in both, usual and map files. This commit also includes adding new articles (3.md and 4.md) and updating content in existing articles (1.md and 2.md). The index.html was also updated to reflect the name change of the main JavaScript file.