:sparkles: Remove multiple articles and related components
The committed changes show the removal of a number of Markdown articles and related TypeScript files, as well as the addition of individual blog pages in the `src/pages` directory. This represents a shift in the project structure and organization, potentially moving from a single articles section to individual blog page components.